Correct tile can no longer be on the same spot twice in succession, changed the game layout

+++ app/src/main/java/com/example/app1/GameActivity.kt
@@ -60,7 +60,7 @@ class GameActivity : AppCompatActivity(){
         transaction.replace(R.id.container, FragmentWin(), "fragment")
         transaction.commit()
 
-        score = (timeLimitSeconds - time.toInt()) * Math.pow(level.toDouble(), 2.0).toInt()
+        score = ((timeLimitSeconds - time) * Math.pow(level.toDouble(), 2.0)).toInt()
         Log.e("GameActivity", "Time: " +  time + ", Score: " + score)
 
         addHighscore(Highscore(currentPlayer, score, time, currentDifficulty))
